Bug 1433158 - Update libvpx to 1.7.0; r=jya
Reviewers: jya

Tags: #secure-revision

Bug #: 1433158

Differential Revision: https://phabricator.services.mozilla.com/D2940

+++ media/libvpx/libvpx/vpx_ports/vpx_timer.h
@@ -83,7 +83,7 @@ static INLINE int64_t vpx_usec_timer_elapsed(struct vpx_usec_timer *t) {
   struct timeval diff;
 
   timersub(&t->end, &t->begin, &diff);
-  return diff.tv_sec * 1000000 + diff.tv_usec;
+  return (int64_t)diff.tv_sec * 1000000 + diff.tv_usec;
 #endif
 }
